requirements.txt
streamlit
pillow
torch
tensorflow
transformers
numpy